Move diff components to diff subdirectory
