+++ web/src/components/diff/DiffStat.scss
