+++ web/src/repo/commit/RepositoryCommitPage.tsx
@@ -25,8 +25,8 @@ import { WebHoverOverlay } from '../../components/shared'
 import { eventLogger, EventLoggerProps } from '../../tracking/eventLogger'
 import { GitCommitNode } from '../commits/GitCommitNode'
 import { gitCommitFragment } from '../commits/RepositoryCommitsPage'
-import { FileDiffConnection } from '../../components/FileDiffConnection'
-import { FileDiffNode } from '../../components/FileDiffNode'
+import { FileDiffConnection } from '../../components/diff/FileDiffConnection'
+import { FileDiffNode } from '../../components/diff/FileDiffNode'
 import { queryRepositoryComparisonFileDiffs } from '../compare/RepositoryCompareDiffPage'
 import { ThemeProps } from '../../../../shared/src/theme'
 import { ErrorAlert } from '../../components/alerts'
