+++ web/src/repo/compare/RepositoryCompareDiffPage.tsx
@@ -12,8 +12,8 @@ import { PlatformContextProps } from '../../../../shared/src/platform/context'
 import { createAggregateError } from '../../../../shared/src/util/errors'
 import { FileSpec, RepoSpec, ResolvedRevSpec, RevSpec } from '../../../../shared/src/util/url'
 import { queryGraphQL } from '../../backend/graphql'
-import { FileDiffConnection } from '../../components/FileDiffConnection'
-import { FileDiffNode } from '../../components/FileDiffNode'
+import { FileDiffConnection } from '../../components/diff/FileDiffConnection'
+import { FileDiffNode } from '../../components/diff/FileDiffNode'
 import { RepositoryCompareAreaPageProps } from './RepositoryCompareArea'
 import { ThemeProps } from '../../../../shared/src/theme'
 import { FileDiffFields, FileDiffHunkRangeFields, DiffStatFields } from '../../backend/diff'
